services/crop_health_and_water_stress_detection/app/model.py
from PIL import Image
import numpy as np
from typing import Dict

class CropStressModel:
    def __init__(self):
        self.loaded = False
        # placeholder for actual model object
        self.model = None

    def load_model(self):
        # replace this with real model loading (torch/keras/tensorflow)
        self.loaded = True

    def preprocess(self, img: Image.Image) -> np.ndarray:
        arr = np.array(img).astype(np.float32)
        # simple resize or normalization can go here if needed
        return arr

    def predict(self, img: Image.Image) -> Dict:
        if not self.loaded:
            self.load_model()
        arr = self.preprocess(img)
        mean = arr.mean()
        # very basic heuristic placeholder
        label = "healthy" if mean > 100 else "water_stress"
        confidence = float(min(max((abs(mean - 100) / 100), 0.01), 0.99))
        return {"label": label, "confidence": confidence}